refactor: :recycle: Unify code for`send_chunks_to_provider`
