+++ client/src/tasks/mod.rs
@@ -14,5 +14,8 @@ pub mod msp_move_bucket;
 pub mod msp_retry_bucket_move;
 pub mod msp_stop_storing_insolvent_user;
 pub mod msp_upload_file;
+pub mod shared {
+    pub mod chunk_uploader;
+}
 pub mod sp_slash_provider;
 pub mod user_sends_file;
